tests/wpt/tests/fedcm/fedcm-after-abort.https.html

<!DOCTYPE html>
<title>Federated Credential Management API network request tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {request_options_with_mediation_required,
        fedcm_test,
        fedcm_get_and_select_first_account} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  let controller = new AbortController();
  let test_options = request_options_with_mediation_required();
  test_options.signal = controller.signal;
  const first_cred = fedcm_get_and_select_first_account(t, test_options);
  controller.abort();
  await promise_rejects_dom(t, 'AbortError', first_cred);

  const second_cred = await fedcm_get_and_select_first_account(t, request_options_with_mediation_required());
  assert_equals(second_cred.token, "token");
}, "Get after abort should work");
</script>
